Move descriptor helpers out to the definition of descriptors
Might consider moving these to instance methods.
