+++ src/descriptors.js
@@ -115,3 +115,36 @@ export class AbstractJoinedDescriptor extends Descriptor {
     return false;
   }
 }
+
+export function cloneDescriptor(d: void | PropertyDescriptor): void | PropertyDescriptor {
+  if (d === undefined) return undefined;
+  return new PropertyDescriptor(d);
+}
+
+// does not check if the contents of value properties are the same
+export function equalDescriptors(d1: PropertyDescriptor, d2: PropertyDescriptor): boolean {
+  if (d1.hasOwnProperty("writable")) {
+    if (!d2.hasOwnProperty("writable")) return false;
+    if (d1.writable !== d2.writable) return false;
+  }
+  if (d1.hasOwnProperty("enumerable")) {
+    if (!d2.hasOwnProperty("enumerable")) return false;
+    if (d1.enumerable !== d2.enumerable) return false;
+  }
+  if (d1.hasOwnProperty("configurable")) {
+    if (!d2.hasOwnProperty("configurable")) return false;
+    if (d1.configurable !== d2.configurable) return false;
+  }
+  if (d1.hasOwnProperty("value")) {
+    if (!d2.hasOwnProperty("value")) return false;
+  }
+  if (d1.hasOwnProperty("get")) {
+    if (!d2.hasOwnProperty("get")) return false;
+    if (d1.get !== d2.get) return false;
+  }
+  if (d1.hasOwnProperty("set")) {
+    if (!d2.hasOwnProperty("set")) return false;
+    if (d1.set !== d2.set) return false;
+  }
+  return true;
+}
